Restructure project into modular components
