+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body class="bg-slate-900">
     <div id="root"></div>
-    <script type="module" src="./src/main.jsx"></script>
+    <script type="module" src="./src/index.js"></script>
   </body>
 </html>
