+++ src/components/DashboardHeader.jsx
@@ -1,0 +1,10 @@
+export default function DashboardHeader() {
+  return (
+    <div className="text-center mb-8 lg:mb-12 mt-12 lg:mt-0">
+      <h1 className="text-3xl lg:text-5xl font-bold text-white mb-4 bg-gradient-to-r from-blue-400 to-purple-400 bg-clip-text text-transparent">
+        Dashboard
+      </h1>
+      <p className="text-slate-300 text-base lg:text-lg">Professional watch market analysis and portfolio management</p>
+    </div>
+  );
+}
