+++ src/data/brands.js
@@ -1,0 +1,10 @@
+export const brands = [
+  'Rolex',
+  'Patek Philippe',
+  'Audemars Piguet',
+  'Omega',
+  'Cartier',
+  'Tudor',
+  'Breitling',
+  'TAG Heuer'
+];
