+++ src/data/conditions.js
@@ -1,0 +1,1 @@
+export const conditions = ['New', 'Like New', 'Excellent', 'Very Good', 'Good', 'Fair'];
