+++ src/data/faqArticles.js
@@ -1,0 +1,38 @@
+export const faqArticles = [
+  {
+    id: 1,
+    title: "How to accurately price your Rolex Submariner",
+    category: "Pricing",
+    content: "Learn the key factors that determine Submariner values including year, condition, box & papers, and market trends."
+  },
+  {
+    id: 2,
+    title: "Understanding watch conditions: New vs Like New vs Excellent",
+    category: "Condition Guide",
+    content: "Detailed guide on how to properly assess and categorize your watch's condition for accurate pricing."
+  },
+  {
+    id: 3,
+    title: "Best platforms for selling luxury watches",
+    category: "Selling",
+    content: "Compare eBay, Chrono24, and other platforms to choose the best one for your watch type and budget."
+  },
+  {
+    id: 4,
+    title: "How to use the Analytics feature effectively",
+    category: "Features",
+    content: "Maximize your market research using our analytics dashboard to track price trends across different watches."
+  },
+  {
+    id: 5,
+    title: "Vintage watch pricing: What makes them valuable",
+    category: "Vintage",
+    content: "Understanding the factors that drive vintage watch prices including rarity, provenance, and historical significance."
+  },
+  {
+    id: 6,
+    title: "Creating compelling watch listings with AI",
+    category: "AI Tools",
+    content: "How to use our Library feature to generate professional titles and descriptions that sell."
+  }
+];
