+++ src/data/models.js
@@ -1,0 +1,10 @@
+export const modelsByBrand = {
+  'Rolex': ['Submariner', 'Daytona', 'Datejust', 'GMT-Master II', 'Explorer', 'Sea-Dweller'],
+  'Patek Philippe': ['Nautilus', 'Aquanaut', 'Calatrava', 'Complications', 'Grand Complications'],
+  'Audemars Piguet': ['Royal Oak', 'Royal Oak Offshore', 'Millenary', 'Jules Audemars'],
+  'Omega': ['Speedmaster', 'Seamaster', 'Constellation', 'De Ville'],
+  'Cartier': ['Tank', 'Santos', 'Ballon Bleu', 'Panthere'],
+  'Tudor': ['Black Bay', 'Pelagos', 'GMT', 'Royal'],
+  'Breitling': ['Navitimer', 'Superocean', 'Avenger', 'Premier'],
+  'TAG Heuer': ['Carrera', 'Monaco', 'Aquaracer', 'Formula 1']
+};
