+++ src/data/platforms.js
@@ -1,0 +1,1 @@
+export const platforms = ['eBay', 'Chrono24', 'FB Marketplace', 'Crown & Caliber', 'Hodinkee Shop', 'WatchStation'];
