+++ src/data/references.js
@@ -1,0 +1,37 @@
+export const referencesByModel = {
+  'Submariner': ['126610LN', '126610LV', '124060', '116610LN', '116610LV', '114060', '16610', '5513', '1680'],
+  'Daytona': ['116500LN', '116503', '116508', '116520', '116523', '116528', '6239', '6241', '6263', '6265'],
+  'Datejust': ['126234', '126200', '126334', '116234', '116200', '16234', '16200', '1601', '1603'],
+  'GMT-Master II': ['126710BLNR', '126710BLRO', '116710BLNR', '116710LN', '16710', '1675'],
+  'Explorer': ['214270', '114270', '1016', '6350', '224270', '124270'],
+  'Sea-Dweller': ['126600', '116600', '16600', '1665', '126603'],
+  'Nautilus': ['5711/1A', '5712/1A', '5726/1A', '5980/1A', '5711/1R', '5712/1R', '5990/1A'],
+  'Aquanaut': ['5167/1A', '5168/G', '5164/A', '5065/A', '5968/A'],
+  'Calatrava': ['5196', '5227', '5296', '5119', '5120', '5127', '5153'],
+  'Complications': ['5205', '5146', '5524', '5396', '5205G', '5235'],
+  'Grand Complications': ['5970', '5204', '5208', '6104', '5002', '5104'],
+  'Royal Oak': ['15202ST', '15400ST', '26331ST', '26470ST', '15500ST', '26320ST', '15300ST'],
+  'Royal Oak Offshore': ['26400SO', '26470OR', '26401RO', '25940SK', '26703ST', '26420SO'],
+  'Millenary': ['77247BC', '77301ST', '15320OR', '77315OR'],
+  'Jules Audemars': ['15180OR', '26100OR', '15056OR', '26390OR'],
+  'Speedmaster': ['310.30.42.50.01.001', '311.30.42.30.01.005', '304.30.44.52.01.001', '105.012', '145.022'],
+  'Seamaster': ['210.30.42.20.01.001', '212.30.41.20.01.003', '233.30.41.21.01.001', '2531.80.00'],
+  'Constellation': ['131.20.39.20.02.001', '123.20.35.20.02.001', '131.25.39.20.52.001'],
+  'De Ville': ['424.23.40.20.02.001', '431.30.41.21.01.001', '424.10.40.20.02.001'],
+  'Tank': ['W5200003', 'W5200004', 'W1529856', 'W1529755', 'WSTA0018', 'WSTA0017'],
+  'Santos': ['WSSA0009', 'WSSA0010', 'W2SA0006', 'W20098D6', 'WGSA0007'],
+  'Ballon Bleu': ['W69012Z4', 'W6920046', 'W2BB0023', 'W69016Z4'],
+  'Panthere': ['W25032P5', 'W25054P5', 'WGPN0006', 'WSPN0006'],
+  'Black Bay': ['79230N', '79230R', '79230B', '79540', '79030N', '79220N'],
+  'Pelagos': ['25600TN', '25600TB', '25407N', '25500TN'],
+  'GMT': ['79830RB', '79830WB', '79833MN'],
+  'Royal': ['28500', '28400', '28300', '28600'],
+  'Navitimer': ['A23322121B2A1', 'A17326211B1A1', 'AB0121211B1A1', 'A13324121B1A1'],
+  'Superocean': ['A17367D71A1A1', 'AB2010121B1A1', 'A1736402BA29'],
+  'Avenger': ['A13385101B1A1', 'A32390111B1A1', 'E13383101B1A1'],
+  'Premier': ['AB0118221B1A1', 'A13315351B1A1', 'A40635161A1A1'],
+  'Carrera': ['CBN2A1B.BA0643', 'CAR201T.BA0766', 'CBM2112.BA0651', 'CV201AH.BA0725'],
+  'Monaco': ['CAW211P.BA0780', 'CBL2111.BA0644', 'CAW211R.FC6401'],
+  'Aquaracer': ['WAY208A.BA0928', 'WAY111A.BA0928', 'CAY211A.BA0927'],
+  'Formula 1': ['CAZ1014.BA0842', 'WAZ1112.BA0875', 'CAZ101E.BA0842']
+};
