+++ src/pages/Portfolio.jsx
@@ -1,0 +1,7 @@
+export default function Portfolio() {
+  return (
+    <div className="text-center py-24 text-slate-300 text-lg">
+      Portfolio Tracking coming soon.
+    </div>
+  );
+}
