+++ src/utils/formatWatchKey.js
@@ -1,0 +1,4 @@
+export default function formatWatchKey(brand, model, reference) {
+  if (!brand || !model) return null;
+  return `${brand}-${model}-${reference || 'all'}`;
+}
